backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
dj-database-url==0.5.0
psycopg2-binary==2.8.6
django-jsonfield-backport==1.0.4
gunicorn==20.1.0